move app.js to source
